user card testing component
