+++ vite.config.ts
@@ -25,6 +25,7 @@ export default defineConfig(({ mode }) => {
         "@": path.resolve(__dirname, "./src"),
         react: path.resolve("./node_modules/react"),
         "react-dom": path.resolve("./node_modules/react-dom"),
+        '@kit': path.resolve(__dirname, '../sn-react-shadcn/src'), //- uncomment when using npm link
         "sn-shadcn-kit": path.resolve(__dirname, "../sn-react-shadcn/src"), //- uncomment when using npm link
       },
     },
